test(reference): consolidate MaxUnpool evaluator tests
